+++ src/main/java/com/example/momo/global/config/WebSocketConfig.java
@@ -1,9 +1,11 @@
-package com.example.momo.global.websocket;
+package com.example.momo.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
 import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
+
+import com.example.momo.global.socket.WebSocketHandler;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
